Move newtreeannotator to treeannotator (the default)
