+++ src/dr/app/tools/treeannotator/BEASTTreesImporter.java
@@ -25,7 +25,7 @@
  *
  */
 
-package dr.app.tools.newtreeannotator;
+package dr.app.tools.treeannotator;
 
 import dr.evolution.io.Importer;
 import dr.evolution.io.TreeImporter;
